+++ src-tauri/src/protocol/responses/usage.rs
@@ -1,0 +1,35 @@
+use serde_json::Value;
+
+/// Parse usage from OpenAI SSE chunk (reuses logic from handlers).
+pub fn parse_usage_from_sse_chunk(text: &str) -> Option<(i64, i64, i64)> {
+    for line in text.lines() {
+        let trimmed = line.trim();
+        if !trimmed.starts_with("data:") {
+            continue;
+        }
+        let data_str = trimmed.trim_start_matches("data:").trim();
+        if data_str == "[DONE]" || data_str.is_empty() {
+            continue;
+        }
+        if let Ok(json) = serde_json::from_str::<Value>(data_str) {
+            if let Some(usage) = json.get("usage") {
+                let prompt = usage
+                    .get("prompt_tokens")
+                    .and_then(|v| v.as_i64())
+                    .unwrap_or(0);
+                let completion = usage
+                    .get("completion_tokens")
+                    .and_then(|v| v.as_i64())
+                    .unwrap_or(0);
+                let total = usage
+                    .get("total_tokens")
+                    .and_then(|v| v.as_i64())
+                    .unwrap_or(0);
+                if total > 0 || prompt > 0 || completion > 0 {
+                    return Some((prompt, completion, total));
+                }
+            }
+        }
+    }
+    None
+}
